Rework interfaces
Added a facade for interacting with docker with a well defined
interface, simplifying the rest of the code. Changed the terraform
and config moudle interfaces to be more consistent and to have more
explicit cleanup of resources.

This should fix the intermittent test failures we were having in
Travis and GitHub Actions - these were caused by a race condition
between output arriving from docker and the code that tested it
running. With the new interfaces it's easier to be confident that
the container will run to completion before the code that tests it
runs.