Change the way diagnostics are collected in the tests
Previously the test containers would put diagnostic information picked up by the
tests onto stderr. I think this was what was causing the intermittent (but pretty
consistent) test failures in GitHub Actions and Travis, so this commit switches
that so most of this is written to a file within a debug volume mapped to the
container. Fingers crossed.