Add release requirements as an interface between the release and config containers
The provides a mechanism for the release contianer to communicate its requirements
to the config container - e.g. "I need a lambda bucket". The release container is
invoked with a single "requirements" parameter and it needs to output a JSON
document containing a map describing its requirements. There is a convention around
what this map should contain (an "env" key with a list of environment variable names),
but cdflow itself is agnostic about this - it just enforces that it must be a map
so the interface is extensible.