Generate the backend configuration if it doesn't exist
We're generating a partial backend configuration (https://www.terraform.io/docs/backends/config.html#partial-configuration)
with the backend type returned from the config container. This gives a user the option to ignore it so that it is
regenerated each time or to commit it, leaving the door open to them providing their own config values if necessary. For
the majority of projects this will be generated on a CI server and the developer will not need to care about it.